+++ Daleks/src/main/java/com/javable/daleks/Settings.java
@@ -1,19 +1,19 @@
 package com.javable.daleks;
 
-public class Settings {
+public class Settings  {
     // stałe programu
     public static final int
             WindowHeight = 600,
             WindowWidth = 1000;
 
     public static final String
-        MainView = "views/main-view.fxml",
-        GameOverView = "views/game-over-view.fxml",
-        PlayerImage = "Daleks/src/main/resources/com/javable/daleks/images/player.jpg",
-        DalekImage = "Daleks/src/main/resources/com/javable/daleks/images/dalek.png",
+            MainView = "views/main-view.fxml",
+            GameOverView = "views/game-over-view.fxml",
+            PlayerImage = "src/main/resources/com/javable/daleks/images/player.jpg",
+            DalekImage = "src/main/resources/com/javable/daleks/images/dalek.png",
 
-        ScrapImage = "Daleks/src/main/resources/com/javable/daleks/images/scrap.jpg",
-        EmptyImage = "Daleks/src/main/resources/com/javable/daleks/images/empty.jpg";
+            ScrapImage = "src/main/resources/com/javable/daleks/images/scrap.jpg",
+            EmptyImage = "src/main/resources/com/javable/daleks/images/empty.jpg";
 
     // ustawienia
     public static int