+++ Daleks/src/main/java/com/javable/daleks/controllers/GameOverController.java
@@ -2,7 +2,6 @@ package com.javable.daleks.controllers;
 
 import com.javable.daleks.DaleksApp;
 import com.javable.daleks.Settings;
-import com.javable.daleks.interfaces.IController;
 import com.javable.daleks.interfaces.IControllerFxmlBased;
 import javafx.fxml.FXML;
 import javafx.scene.control.Label;