+++ Daleks/src/main/java/com/javable/daleks/controllers/GridManager.java
@@ -3,11 +3,14 @@ package com.javable.daleks.controllers;
 import com.javable.daleks.Settings;
 import com.javable.daleks.enums.EDirection;
 import com.javable.daleks.enums.EObjectType;
+import com.javable.daleks.logic.ImageLoader;
 import com.javable.daleks.models.GameMap;
 import com.javable.daleks.models.Position;
 import com.javable.daleks.models.objects.Dalek;
 import com.javable.daleks.models.objects.Scrap;
 import javafx.geometry.HPos;
+import javafx.geometry.Insets;
+import javafx.geometry.Pos;
 import javafx.scene.effect.ColorAdjust;
 import javafx.scene.effect.Effect;
 import javafx.scene.image.ImageView;
@@ -15,80 +18,76 @@ import javafx.scene.layout.ColumnConstraints;
 import javafx.scene.layout.GridPane;
 import javafx.scene.layout.RowConstraints;
 
+import java.io.FileNotFoundException;
+
 public class GridManager {
 
-    private final com.javable.daleks.logic.ImageLoader ImageLoader;
+    private final ImageLoader imageLoader = new ImageLoader();
     private final ImageView[][] cells;
     private final GameMap map;
-    private final GridPane GameGrid;
+    private final GridPane gameGrid;
+    private final int gridSize;
 
 
-    public GridManager(com.javable.daleks.logic.ImageLoader imageLoader, ImageView[][] cells, GameMap gameMap, GridPane gameGrid) {
-        ImageLoader = imageLoader;
-        this.cells = cells;
+    public GridManager(GridPane gameGrid, GameMap gameMap) throws FileNotFoundException {
+        this.gameGrid = gameGrid;
+        gameGrid.setAlignment(Pos.CENTER);
+        gameGrid.setPadding(new Insets(10, 10, 10, 10));
+        gameGrid.setGridLinesVisible(true);
+
+        cells = new ImageView[gameMap.gridCount][];
+        gridSize = (Settings.WindowHeight-100)/ gameMap.gridCount;
         map = gameMap;
-        GameGrid = gameGrid;
+
+        this.initialize();
     }
 
-//    public void FrameUpdate(EDirection direction) {
-//        /*
-//        The next time we update the grid we only care about what has changed
-//         */
-//        List<Position> updatedCells = map.NextFrame(direction);
-//
-//        for (Position cell : updatedCells) {
-//            EObjectType objectAtOldPosition = map.GetObjectAtCell(cell);
-//
-//            Image image = ImageLoader.ToImage(objectAtOldPosition);
-//
-//            cells[cell.x][cell.y].setImage(image);
-//        }
-//    }
-    public void repaint(){
+    public void repaint() {
         clear();
-        cells[map.Player.position.x][map.Player.position.y].setImage(ImageLoader.ToImage(EObjectType.Player));
+        cells[map.player.position.x][map.player.position.y].setImage(imageLoader.getImage(EObjectType.Player));
 
-        for (EDirection direction: EDirection.values()) {
-            Position playerPosition = map.Player.position;
-            Position newPosition = Position.Move(playerPosition, direction);
-            if(GameMap.IsInBounds(newPosition) && map.GetObjectAtCell(newPosition).isEmpty()){
+        for (EDirection direction : EDirection.values()) {
+            Position playerPosition = map.player.position;
+            Position newPosition = playerPosition.add(direction.toVector());
+            if (map.isInBounds(newPosition) && map.getObjectAtCell(newPosition).isEmpty()) {
                 ImageView current = cells[newPosition.x][newPosition.y];
-                Effect effect = new ColorAdjust(1,1,0.5,1);
+                Effect effect = new ColorAdjust(1, 1, 0.5, 1);
                 current.setEffect(effect);
             }
 
         }
-        for(Dalek dalek:map.Daleks)
-        {
-            cells[dalek.position.x][dalek.position.y].setImage(ImageLoader.ToImage(EObjectType.Dalek));
+        for (Dalek dalek : map.daleks) {
+            cells[dalek.position.x][dalek.position.y].setImage(imageLoader.getImage(EObjectType.Dalek));
         }
-        for(Scrap scrap : map.Scrap){
-            cells[scrap.position.x][scrap.position.y].setImage(ImageLoader.ToImage(EObjectType.Scrap));
+        for (Scrap scrap : map.scraps) {
+            cells[scrap.position.x][scrap.position.y].setImage(imageLoader.getImage(EObjectType.Scrap));
         }
 
     }
-    public void clear(){
-        for (int i = 0; i < Settings.GridCount; i++) {
-            for (int j = 0; j < Settings.GridCount; j++) {
-                cells[i][j].setImage(ImageLoader.ToImage(EObjectType.Empty));
+
+    public void clear() {
+        for (int i = 0; i < map.gridCount; i++) {
+            for (int j = 0; j < map.gridCount; j++) {
+                cells[i][j].setImage(imageLoader.getImage(EObjectType.Empty));
                 cells[i][j].setEffect(null);
             }
         }
     }
-    public void Initialize() {
 
-        for (int i = 0; i < Settings.GridCount; i++) {
-            GameGrid.getColumnConstraints().add(new ColumnConstraints(Settings.GridSize));
-            cells[i] = new ImageView[Settings.GridCount];
+    public void initialize() {
+
+        for (int i = 0; i < map.gridCount; i++) {
+            gameGrid.getColumnConstraints().add(new ColumnConstraints(gridSize));
+            cells[i] = new ImageView[map.gridCount];
 
 
-            for (int j = 0; j < Settings.GridCount; j++) {
+            for (int j = 0; j < map.gridCount; j++) {
                 if (i == 0)
-                    GameGrid.getRowConstraints().add(new RowConstraints(Settings.GridSize));
-                cells[i][j] = new ImageView(ImageLoader.ToImage(EObjectType.Empty));
-                cells[i][j].setFitHeight(40);
+                    gameGrid.getRowConstraints().add(new RowConstraints(gridSize));
+                cells[i][j] = new ImageView(imageLoader.getImage(EObjectType.Empty));
+                cells[i][j].setFitHeight(gridSize-5);
                 cells[i][j].setPreserveRatio(true);
-                GameGrid.add(cells[i][j], i, j);
+                gameGrid.add(cells[i][j], i, j);
                 GridPane.setHalignment(cells[i][j], HPos.CENTER);
             }
         }