+++ Daleks/src/main/java/com/javable/daleks/controllers/MainController.java
@@ -1,21 +1,33 @@
 package com.javable.daleks.controllers;
 
+import com.google.inject.Guice;
+import com.google.inject.Injector;
 import com.javable.daleks.Settings;
-import com.javable.daleks.interfaces.IController;
 import com.javable.daleks.interfaces.IControllerFxmlBased;
+import com.javable.daleks.models.GameMapSettings;
 import javafx.fxml.FXML;
-import javafx.scene.control.Label;
+import javafx.scene.control.TextField;
 
 import java.io.IOException;
 
 public class MainController implements IControllerFxmlBased {
     @FXML
-    private Label welcomeText;
+    private TextField daleks_count_input, map_size_input;
+
+    private GameMapSettings parseInput() {
+        int daleks_count = Integer.parseInt(daleks_count_input.getText());
+        int map_size = Integer.parseInt(map_size_input.getText());
+        return new GameMapSettings(
+                map_size,
+                daleks_count
+        );
+    }
 
     @FXML
     protected void NewGameBtn() throws IOException {
-        //welcomeText.setText("Welcome to JavaFX Application!");
-        GameController gameController = new GameController();
+        GameMapSettings settings = parseInput();
+        Injector injector = Guice.createInjector(settings);
+        GameController gameController = injector.getInstance(GameController.class);
         gameController.InitView();
     }