+++ Daleks/src/main/java/com/javable/daleks/interfaces/IController.java
@@ -1,7 +1,5 @@
 package com.javable.daleks.interfaces;
 
-import com.javable.daleks.logic.ViewManager;
-
 import java.io.IOException;
 
 public interface IController {