+++ Daleks/src/main/java/com/javable/daleks/logic/CollisionHandlerVisitor.java
@@ -1,11 +1,13 @@
 package com.javable.daleks.logic;
 
+import com.javable.daleks.Settings;
 import com.javable.daleks.models.GameMap;
-import com.javable.daleks.models.Position;
 import com.javable.daleks.models.objects.Dalek;
 import com.javable.daleks.models.objects.ObjectBase;
 import com.javable.daleks.models.objects.Player;
 import com.javable.daleks.models.objects.Scrap;
+
+import java.io.IOException;
 
 public class CollisionHandlerVisitor {
 
@@ -22,19 +24,26 @@ public class CollisionHandlerVisitor {
         No matter what a player collides with he will die
          */
         player.kill();
+        try {
+            ViewManager.SetScene(Settings.GameOverView);
+        } catch (IOException e) {
+            throw new RuntimeException(e);
+        }
     }
+
     public void handleDalekCollision(Dalek dalek, ObjectBase other, boolean inWalk) {
         /*
         Remove self from map if other is not player.
         Add scrap to map if needed.
          */
-        if(!(other instanceof Player)) {
+        if (!(other instanceof Player)) {
             map.removeDalek(dalek);
-            if(inWalk)
-                map.AddScrap(new Scrap(dalek.position));
+            if (inWalk)
+                map.addScrap(new Scrap(dalek.position));
         }
 
     }
+
     public void handleScrapCollision(Scrap scrap, ObjectBase other, boolean inWalk) {
         return;
     }