+++ Daleks/src/main/java/com/javable/daleks/logic/InputHandler.java
@@ -9,17 +9,20 @@ import javafx.scene.layout.GridPane;
 public class InputHandler {
 
     MoveHandler moveHandler;
-    public InputHandler(MoveHandler moveHandler){
+
+    public InputHandler(MoveHandler moveHandler) {
         this.moveHandler = moveHandler;
     }
+
     public void clickGrid(GridManager gridManager, GridPane GameGrid, EventTarget target) {
         Node clickedNode = (Node) target;
         if (clickedNode != GameGrid) {
             // click on descendant node
             Integer colIndex = GridPane.getColumnIndex(clickedNode);
             Integer rowIndex = GridPane.getRowIndex(clickedNode);
-            moveHandler.MovePlayer(new Position(colIndex, rowIndex));
+            moveHandler.movePlayer(new Position(colIndex, rowIndex));
         }
     }
+
 
 }