+++ Daleks/src/main/java/com/javable/daleks/models/objects/Dalek.java
@@ -5,13 +5,12 @@ import com.javable.daleks.enums.EObjectType;
 import com.javable.daleks.logic.CollisionHandlerVisitor;
 import com.javable.daleks.models.Position;
 
-public class Dalek extends ObjectBase{
+public class Dalek extends ObjectBase {
 
     public Dalek(Position position) {
         super(position, EObjectType.Dalek);
     }
 
-    @Override
     public boolean canMove(EDirection direction) {
         return true;
     }