+++ Daleks/src/main/java/com/javable/daleks/models/objects/ObjectBase.java
@@ -2,16 +2,12 @@ package com.javable.daleks.models.objects;
 
 import com.javable.daleks.enums.EDirection;
 import com.javable.daleks.enums.EObjectType;
-import com.javable.daleks.interfaces.IMovable;
 import com.javable.daleks.logic.CollisionHandlerVisitor;
 import com.javable.daleks.models.Position;
-import kotlin.NotImplementedError;
 
-import java.util.Vector;
-
-public abstract class ObjectBase implements IMovable {
-    public Position position;
+public abstract class ObjectBase {
     public final EObjectType objectType;
+    public Position position;
 
     public ObjectBase(Position position, EObjectType type) {
         this.position = position;
@@ -19,7 +15,7 @@ public abstract class ObjectBase implements IMovable {
     }
 
     public void move(EDirection direction) {
-        position.add(Position.ToVector(direction));
+        position = position.add(direction.toVector());
     }
 
     public abstract void accept(CollisionHandlerVisitor visitor, ObjectBase other, boolean inWalk);